# Use regexp.MustCompile at initialize 