--- a/detector.go
+++ b/detector.go
@@ -5,16 +5,19 @@ import (
 	"strings"
 )
 
+var (
+	crawlerReg   = regexp.MustCompile(CombineRegexp(CrawlersList()))
+	exclusionReg = regexp.MustCompile(CombineRegexp(ExclusionsList()))
+)
+
 // CrawlerDetector is crawler detector structure
 type CrawlerDetector struct {
-	Crawlers   []string
-	Exclusions []string
-	Matched    []string
+	Matched []string
 }
 
 // New returns a new initialized CrawlerDetector
 func New() *CrawlerDetector {
-	return &CrawlerDetector{CrawlersList(), ExclusionsList(), []string{}}
+	return &CrawlerDetector{Matched: []string{}}
 }
 
 // IsCrawler is detect crawlers/spiders/bots by user agent
@@ -23,8 +26,7 @@ func (cd *CrawlerDetector) IsCrawler(userAgent string) bool {
 		return false
 	}
 
-	cReg := regexp.MustCompile(cd.CombineRegexp(cd.Crawlers))
-	cd.Matched = cReg.FindAllString(userAgent, -1)
+	cd.Matched = crawlerReg.FindAllString(userAgent, -1)
 
 	if len(cd.Matched) != 0 {
 		return true
@@ -35,8 +37,7 @@ func (cd *CrawlerDetector) IsCrawler(userAgent string) bool {
 
 // IsExclusion is detect exclusion from user agent
 func (cd *CrawlerDetector) IsExclusion(userAgent string) bool {
-	eReg := regexp.MustCompile(cd.CombineRegexp(cd.Exclusions))
-	isExclusion := eReg.ReplaceAllString(userAgent, "")
+	isExclusion := exclusionReg.ReplaceAllString(userAgent, "")
 
 	if len(isExclusion) == 0 {
 		return true
@@ -46,19 +47,19 @@ func (cd *CrawlerDetector) IsExclusion(userAgent string) bool {
 }
 
 // CombineRegexp is build regex from givement patterns list
-func (cd *CrawlerDetector) CombineRegexp(patterns []string) string {
+func CombineRegexp(patterns []string) string {
 	return "(" + strings.Join(patterns, "|") + ")"
 }
 
 // SetCrawlers is setter for custom crawlers list
 func (cd *CrawlerDetector) SetCrawlers(list []string) *CrawlerDetector {
-	cd.Crawlers = list
+	crawlerReg = regexp.MustCompile(CombineRegexp(list))
 	return cd
 }
 
 // SetExclusions is setter for custom exclusions list
 func (cd *CrawlerDetector) SetExclusions(list []string) *CrawlerDetector {
-	cd.Exclusions = list
+	exclusionReg = regexp.MustCompile(CombineRegexp(list))
 	return cd
 }